# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/auth.LoginResponseData'
      message:
        type: string
      status:
        type: string
    type: object
...
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterRequest:
    properties:
      accountType:
        $ref: '#/definitions/common.AccountType'
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      referralUrl:
        type: string
    required:
    - email
    - firstName
    - name
    - password
    type: object
  auth.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
...
  common.AccountType:
    enum:
    - attendee
    - organizer
    type: string
    x-enum-varnames:
    - AccountTypeAttendee
    - AccountTypeOrganizer
...
  event.StatusRequest:
    properties:
      status:
        type: string
    type: object
  event.UpdateEventImageRequest:
    properties:
      eventId:
        type: string
      imageUrl:
        type: string
    type: object
  event.UpdateEventVideoRequest:
    properties:
      eventId:
        type: string
      videoUrl:
        type: string
    type: object
...
  github_com_gbaski_gbaski-host_internal_event.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      otp_required:
        type: boolean
      status:
        example: success
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gbaski.app
    name: API Support
  description: API documentation for Gbaski Host service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gbaski Host API
  version: "1.0"
paths:
  /auth/available-name/{name}:
    get:
      consumes:
      - application/json
      description: Check if a name is available for registration
      parameters:
      - description: Name to check
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Check if a name is available
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Login user
      tags:
      - auth
  /auth/otp-login:
    post:
      consumes:
      - application/json
      description: Authenticate a user using OTP code sent to their email
      parameters:
      - description: Email address
        in: query
        name: otp_email
        required: true
        type: string
      - description: OTP code (required for verification)
        in: query
        name: otp_code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/auth.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Login user with OTP
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Get a new access token using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/auth.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.Response'
      summary: Register new user
      tags:
      - auth
...
  /events:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of events for the authenticated user
        with optional filtering and search
      parameters:
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get all events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new event for the authenticated user
      parameters:
      - description: Event creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Create new event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event by ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Delete event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Retrieve a single event by its ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event by ID
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update an existing event by ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Event update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.CreateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event
      tags:
      - events
  /events/{id}/stats:
    get:
      consumes:
      - application/json
      description: Retrieve statistics for a specific event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event statistics
      tags:
      - events
  /events/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of an existing event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.StatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event status
      tags:
      - events
  /events/categories:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all available event categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event categories
      tags:
      - events
  /events/image:
    post:
      consumes:
      - application/json
      description: Update the image URL for an event
      parameters:
      - description: Event image update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.UpdateEventImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event image
      tags:
      - events
  /events/report:
    get:
      consumes:
      - application/json
      description: Retrieve a report of all events for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get events report
      tags:
      - events
  /events/video:
    post:
      consumes:
      - application/json
      description: Update the video URL for an event
      parameters:
      - description: Event video update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/event.UpdateEventVideoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Update event video
      tags:
      - events
  /payout/account:
    get:
      consumes:
      - application/json
      description: Retrieve the payout account for the authenticated user
      parameters:
      - default: NGN
        description: Currency code
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_common.Response'
      security:
      - Bearer: []
      summary: Get payout account
      tags:
      - payout
  /payout/banks:
    get:
      consumes:
      - application/json
      description: Retrieve a list of available banks for payout account setup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout banks
      tags:
      - payout
  /payout/change-type:
    post:
      consumes:
      - application/json
      description: Change the payout type for a payout account
      parameters:
      - description: Payout type change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.ChangePayoutTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Change payout type
      tags:
      - payout
  /payout/create-account:
    post:
      consumes:
      - application/json
      description: Create a new payout account for the authenticated user with OTP
        verification
      parameters:
      - description: OTP code for verification
        in: query
        name: otp_code
        type: string
      - description: Payout account creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.CreatePayoutAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Create payout account
      tags:
      - payout
  /payout/history/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve paginated payout history for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      - description: Search query
        in: query
        name: search
        type: string
      - description: Filter parameters
        in: query
        name: filter
        type: string
      - description: Range parameters
        in: query
        name: range
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout history
      tags:
      - payout
  /payout/payment-provider:
    get:
      consumes:
      - application/json
      description: Retrieve payment provider integration details for the authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payment payout provider
      tags:
      - payout
    post:
      consumes:
      - application/json
      description: Create or update payment provider integration for payouts
      parameters:
      - description: Payment payout provider details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payout.CreatePaymentPayoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Create payment payout provider
      tags:
      - payout
  /payout/summary/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve payout summary for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get payout summary
      tags:
      - payout
  /payout/total-payout/{formId}:
    get:
      consumes:
      - application/json
      description: Retrieve total payout amount for a specific form
      parameters:
      - description: Form ID (UUID)
        in: path
        name: formId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payout.Response'
      security:
      - Bearer: []
      summary: Get total payout
      tags:
      - payout
  /registrations/check-in:
    get:
      consumes:
      - application/json
      description: Check in a registration using registration reference and agent
        token
      parameters:
      - description: Registration reference
        in: query
        name: regRef
        required: true
        type: string
      - description: Bearer token (check-in agent token)
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/registration.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/registration.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/registration.Response'
      summary: Check in registration
      tags:
      - registrations
  /registrations/forms:
    get:
      consumes:
      - application/json
      description: Retrieve all registration forms for the authenticated user's events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Get event forms
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Create a new registration form for an event
      parameters:
      - description: Form creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.FormRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_gbaski_gbaski-host_internal_event.Response'
      security:
      - Bearer: []
      summary: Create registration form
      tags:
      - events
...
schemes:
- http
- https
swagger: "2.0"